rework RustPlus socket
